address code smell

+++ streets_utils/streets_tsc_configuration/include/tsc_configuration_state.h
@@ -51,7 +51,7 @@ namespace streets_tsc_configuration
     {
         std::vector<signal_group_configuration> tsc_config_list;
 
-        signal_group_configuration get_signal_group_configuration_by_sg(uint8_t signal_group_id);
+        signal_group_configuration get_signal_group_configuration_by_sg(int signal_group_id);
 
         /**
          * @brief Serialize tsc_configuration_state object to rapidjson::Value for writing as JSON string
