+++ streets_utils/streets_tsc_configuration/src/tsc_configuration_state.cpp
@@ -67,12 +67,12 @@ namespace streets_tsc_configuration
     }
 
     
-    signal_group_configuration tsc_configuration_state::get_signal_group_configuration_by_sg(uint8_t signal_group_id)
+    signal_group_configuration tsc_configuration_state::get_signal_group_configuration_by_sg(int signal_group_id)
     {
         signal_group_configuration result_sg_config;
         for(const auto &sg_config: tsc_config_list)
         {
-            if(sg_config.signal_group_id == signal_group_id)
+            if(static_cast<int>(sg_config.signal_group_id)== signal_group_id)
             {
                 result_sg_config = sg_config;
             }
